typescript * Renamer OppsumeringsElement -> Element
