+++ src/app/element.less
@@ -1,4 +1,4 @@
-.oppsummerings-element {
+.element {
   margin-bottom: 0.25cm;
 
   &__sporsmal {
