+++ src/app/index.less
@@ -1,6 +1,6 @@
 @import "header";
 @import "innsending-tidspunkt";
-@import "oppsummerings-element";
+@import "element";
 @import "bolk";
 @import "bekreftelse";
 @import "soknads-info";
